feat: Enhance Neovim configuration and key mappings
This commit includes several enhancements to the Neovim configuration
and key mappings. It sets the 'magma_image_provider' option in the
Neovim configuration file to use 'kitty' as the image provider. It also
updates the key mappings for tmux navigation and DAP. Additionally, it
adds new plugins for executing code like Jupyter and sets up DAP for
Python. The changes will take effect after Neovim is restarted.