feat(neovim): Refactor LSP setup and add Neovim download in Codespaces
This commit refactors the Language Server Protocol (LSP) setup in the
Neovim configuration file. The setup for 'pyright', 'ltex', and 'lua_ls'
has been moved to a new 'mason-lspconfig' module. This makes the LSP
configuration more modular and easier to manage.

Additionally, this commit adds a new feature to automatically download
and unzip Neovim in GitHub Codespaces. This ensures that Neovim is
readily available in any new Codespace.