feat: Move LSP configuration to mason.lua and update install script
This commit moves the Language Server Protocol (LSP) configuration from
lsp.lua to mason.lua. This change is made to centralize the configuration
and make it easier to manage.

In addition, the install script is updated to create a symbolic link to
Neovim in the /usr/.local/bin directory. This ensures that Neovim is
accessible from any location in the terminal.

Lastly, the TERM environment variable is set to 'xterm' to ensure
compatibility with certain terminal-based applications.