feat(nvim): improve keymap and copilot settings
This commit includes two main changes. First, the keymap for
'<leader>f' in Neovim has been updated to check if the current
directory is a git repository. If it is, it uses the
'telescope_builtin.git_files' function. If not, it uses the
'telescope_builtin.find_files' function. This makes file searching
more context-aware and efficient.

Second, the settings for the Copilot plugin have been updated to
include mappings for resetting the chat. This can be done in both
normal and insert modes by pressing '<C-r>'.