feat: Update diagnostic keymaps and symbols
This commit updates the keymaps for navigating diagnostic messages
in Neovim. The keys for going to the next and previous diagnostic
message have been swapped for better consistency.

Additionally, a new symbol for "hint" diagnostic messages has been
added to the status line configuration. To reduce reliance on hack fonts