feat(nvim): streamline plugin installation and color scheme setup
This commit makes two main changes to the Neovim configuration. First,
it simplifies the plugin installation process by removing the
`is_bootstrap` variable and directly requiring `packer` to sync after
cloning the repository. This ensures that plugins are installed before
loading configs.

Second, it moves the color scheme setup to a more appropriate location
in the file, removing the need for the `termguicolors` option. The
`dracula` color scheme is set and the background for `Normal` and
`NormalFloat` highlight groups is set to none.